Create Equivalence Classes
After testing the upper boundary of 1,000,000, the program was taking a long time to execute. So, I reduced the upper bound to 100,000 instead. And for the Negative, the function is stuck in an infinite loop because the cycle starts with -1, which is below 1 already. This makes the number decrease further from the target number of 1. Everything else worked well.